+++ src/Service/SecretsManager/src/Input/UpdateSecretRequest.php
@@ -191,9 +191,10 @@ final class UpdateSecretRequest extends Input
             throw new InvalidArgument(sprintf('Missing parameter "SecretId" for "%s". The value cannot be null.', __CLASS__));
         }
         $payload['SecretId'] = $v;
-        if (null !== $v = $this->clientRequestToken) {
-            $payload['ClientRequestToken'] = $v;
+        if (null === $v = $this->clientRequestToken) {
+            $v = uuid_create(\UUID_TYPE_RANDOM);
         }
+        $payload['ClientRequestToken'] = $v;
         if (null !== $v = $this->description) {
             $payload['Description'] = $v;
         }