+++ src/Service/SecretsManager/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Fill the optional `clientRequestToken` property with an UUID when value is omitted.
 - AWS enhancement: Documentation updates for SecretsManager.
 
 ## 0.1.0